Bismillah/license_server/requirements.txt
bip_utils
eth_account
fastapi
uvicorn[standard]
slowapi
apscheduler
httpx
supabase
python-dotenv
hypothesis
pytest
pytest-asyncio
